feat: добавление словаря "тип валюты"

+++ src/main/java/ru/pssbd/fonds/repository/CurrencyTypeRepository.java
@@ -4,8 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.pssbd.fonds.entity.CurrencyTypeEntity;
 
-import java.math.BigInteger;
-
 @Repository
-public interface CurrencyTypeRepository extends JpaRepository<CurrencyTypeEntity, BigInteger> {
+public interface CurrencyTypeRepository extends JpaRepository<CurrencyTypeEntity, Short> {
 }
